requirements.txt
# Core dependencies
litellm
streamlit
python-dotenv
asyncio

# HTTP and API communication
aiohttp
certifi
